Fixed units + Made site responsive
Changed px units to rem for fonts to make site more responsive and not override user font settings. Also changed paddign to em so its in porportion to the font sizes.

Added two media queries. One makes the heading text smaller when the screen width gets small enough. When the screen gets even smaller, the left side of the page goes to the top and the right side goes under.